requirements.txt
Flask==3.1.0
gunicorn==20.1.0
firebase-admin==6.6.0
face_recognition==1.3.0
torch==2.6.0+cpu  # CPU-only build; Dockerfile adds the PyTorch CPU wheel index
opencv-python-headless==4.11.0.86
pillow==11.1.0
pillow_heif==0.21.0
mysql-connector-python==9.2.0
numpy==2.0.2
google-cloud-storage==2.10.0
stripe==11.6.0
msal==1.32.0
requests==2.31.0
google-api-python-client==2.109.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
